pyproject.toml
[tool.poetry]
name = "assessment"
version = "0.1.0"
description = ""
authors = ["ArshadQ118 <arshadq118@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.3"
djangorestframework = "3.14.0"
django-cors-headers = "4.1.0"
boto3 = "1.26.158"
redis = "^5.0.3"
celery = "^5.3.6"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
